+++ stm32.h
@@ -56,7 +56,7 @@ char stm32_write_memory  (const stm32_t *stm, uint32_t address, uint8_t data[], 
 char stm32_wunprot_memory(const stm32_t *stm);
 char stm32_readprotect_memory(const stm32_t *stm);
 char stm32_readunprotect_memory(const stm32_t *stm);
-char stm32_erase_memory  (const stm32_t *stm, uint8_t spage, uint8_t pages);
+char stm32_erase_memory  (const stm32_t *stm, uint8_t spage, uint16_t pages);
 char stm32_go            (const stm32_t *stm, uint32_t address);
 char stm32_reset_device  (const stm32_t *stm);
